# optimized to_pil by 2.71x (63.17% improvement)