Input type: <class 'torch.Tensor'>
Valid outputs: True
Original time: 1.249 ms
Optimized time: 0.539 ms
Speedup: 2.32x

Input type: <class 'torch.Tensor'>
Valid outputs: True
Original time: 57.895 ms
Optimized time: 21.326 ms
Speedup: 2.71x

https://colab.research.google.com/drive/18B6fsMMOex0FArRsceI95Cid4h-WEBNN?usp=sharing
Here's the benchmarking and validation script

The optimized version uses in-place operations and converts the entire batch tensor to NumPy at once, minimizing memory usage and costly repeated conversions. It also ensures contiguous memory for faster .numpy() calls, making it significantly faster and more efficient for batch image processing.